# apps/indigo-api-gateway/application/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API gateway for Indigo Server.",
        "title": "Indigo Server",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/v1/auth/google/callback": {
            "post": {
                "description": "Handles the callback from the Google OAuth2 provider to exchange the authorization code for a token, verify the user, and issue access and refresh tokens.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication API"
                ],
                "summary": "Google OAuth2 Callback",
                "parameters": [
                    {
                        "description": "Request body containing the authorization code and state",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/app_interfaces_http_routes_v1_auth_google.GoogleCallbackRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully authenticated and returned tokens",
                        "schema": {
                            "$ref": "#/definitions/app_interfaces_http_routes_v1_auth_google.AccessTokenResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request (e.g., invalid state, missing code, or invalid claims)",
                        "schema": {
                            "$ref": "#/definitions/menlo_ai_indigo-api-gateway_app_interfaces_http_responses.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized (e.g., a user claim is not found or is invalid in the context)",
                        "schema": {
                            "$ref": "#/definitions/menlo_ai_indigo-api-gateway_app_interfaces_http_responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/menlo_ai_indigo-api-gateway_app_interfaces_http_responses.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/auth/google/login": {
            "get": {
                "description": "Redirects the user to the Google OAuth2 authorization page to initiate the login process.",
                "tags": [
                    "Authentication API"
                ],
                "summary": "Google OAuth2 Login",
                "responses": {
                    "200": {
                        "description": "redirect url",
                        "schema": {
                            "$ref": "#/definitions/app_interfaces_http_routes_v1_auth_google.GoogleLoginUrl"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/menlo_ai_indigo-api-gateway_app_interfaces_http_responses.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/v1/conv/models": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves a list of available models that can be used for conversation-aware chat completions. This endpoint provides the same model list as the standard /v1/models endpoint but is specifically designed for conversation-aware chat functionality.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Conversation-aware Chat API"
                ],
                "summary": "List available models for conversation-aware chat",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "$ref": "#/definitions/app_interfaces_http_routes_v1_conv.ModelsResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - missing or invalid authentication",
                        "schema": {
                            "$ref": "#/definitions/menlo_ai_indigo-api-gateway_app_interfaces_http_responses.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/v1/models": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves a list of available models that can be used for chat completions or other tasks.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chat Completions API"
                ],
                "summary": "List available models",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "$ref": "#/definitions/app_interfaces_http_routes_v1.ModelsResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "app_interfaces_http_routes_v1.Model": {
            "type": "object",
            "properties": {
                "created": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "object": {
                    "type": "string"
                },
                "owned_by": {
                    "type": "string"
                }
            }
        },
        "app_interfaces_http_routes_v1.ModelsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/app_interfaces_http_routes_v1.Model"
                    }
                },
                "object": {
                    "type": "string"
                }
            }
        },
...
        "app_interfaces_http_routes_v1_auth_google.AccessTokenResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "expires_in": {
                    "type": "integer"
                },
                "object": {
                    "type": "string"
                }
            }
        },
        "app_interfaces_http_routes_v1_auth_google.GoogleCallbackRequest": {
            "type": "object",
            "required": [
                "code"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                }
            }
        },
        "app_interfaces_http_routes_v1_auth_google.GoogleLoginUrl": {
            "type": "object",
            "properties": {
                "object": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "app_interfaces_http_routes_v1_conv.Model": {
            "type": "object",
            "properties": {
                "created": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "object": {
                    "type": "string"
                },
                "owned_by": {
                    "type": "string"
                }
            }
        },
        "app_interfaces_http_routes_v1_conv.ModelsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/app_interfaces_http_routes_v1_conv.Model"
                    }
                },
                "object": {
                    "type": "string"
                }
            }
        },
...
        "menlo_ai_indigo-api-gateway_app_interfaces_http_responses.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}